requirements.txt
yt-dlp
moviepy
discord.py
fastapi
uvicorn
python-dotenv
requests